move scale to zero to runtime util
